+++ gfx/snes_state.h
@@ -52,7 +52,7 @@ struct snes_tracker_uniform_info
    uint32_t addr;
    enum snes_tracker_type type;
    enum snes_ram_type ram_type;
-   uint8_t mask;
+   unsigned mask;
 };
 
 struct snes_tracker_info
